backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
betterpath==0.2.2
click==8.0.3
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
l==0.11.0
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
vcversioner==2.16.0.0
whitenoise==5.3.0
zope.interface==5.4.0
